sort out the difference between player/color
On the client. This now makes swap work. If you don't get the
distinction between player and color right on both the client
and the server, they will start to disagree about whose turn
it is.

The nice way to do this seems to be to compute the current player's
color from (a) the current player and (b) whether a swap happened.